pyproject.toml
[tool.poetry]
name = "cw-django-email-sender"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
python-dotenv = "^1.0.1"
django-phonenumber-field = "^8.0.0"
psycopg2-binary = "^2.9.9"
phonenumbers = "^8.13.44"
ipython = "^8.27.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
